站点，模型，设备id改成长id，以及模型和设备的批量导入导出

+++ app/model/model/flow_folders.js
@@ -32,7 +32,7 @@ module.exports = app => {
       field: 'component',
     },
     station_id: {
-      type: DataTypes.INTEGER(11).UNSIGNED,
+      type: DataTypes.BIGINT.UNSIGNED,
       allowNull: true,
       defaultValue: null,
       primaryKey: false,
